Use external sagalog-api dependency.
