+++ pom.xml
@@ -56,6 +56,11 @@
             <version>0.1</version>
         </dependency>
         <dependency>
+            <groupId>no.ssb.saga</groupId>
+            <artifactId>sagalog-api</artifactId>
+            <version>0.1</version>
+        </dependency>
+        <dependency>
             <groupId>no.ssb.concurrent.futureselector</groupId>
             <artifactId>future-selector</artifactId>
             <version>0.2</version>
@@ -181,7 +186,7 @@
                 <configuration>
                     <serverId>ossrh</serverId>
                     <nexusUrl>https://oss.sonatype.org/</nexusUrl>
-                    <autoReleaseAfterClose>false</autoReleaseAfterClose>
+                    <autoReleaseAfterClose>true</autoReleaseAfterClose>
                 </configuration>
             </plugin>
             <plugin>
