--- src/main/java/no/ssb/saga/execution/sagalog/SagaLogEntryId.java
@@ -1,4 +1,0 @@
-package no.ssb.saga.execution.sagalog;
-
-public interface SagaLogEntryId {
-}
