--- src/main/java/no/ssb/saga/execution/sagalog/SagaLogEntryType.java
@@ -1,9 +1,0 @@
-package no.ssb.saga.execution.sagalog;
-
-public enum SagaLogEntryType {
-    Start,
-    End,
-    Abort,
-    Comp,
-    Ignore // Entries of this type should always be ignored, may be used internally by the saga-log.
-}
