--- src/main/java/no/ssb/saga/execution/sagalog/SagaLogPool.java
@@ -1,8 +1,0 @@
-package no.ssb.saga.execution.sagalog;
-
-public interface SagaLogPool {
-
-    SagaLog connect(String logId);
-
-    void release(String logId);
-}
