--- src/main/resources/META-INF/services/no.ssb.saga.execution.sagalog.SagaLog
@@ -1,1 +1,0 @@
-no.ssb.saga.execution.sagalog.memory.MemorySagaLog
